requirements.txt
amqp==5.1.1
APScheduler==3.6.3  # Поменяли на версию 3.6.3 для совместимости с python-telegram-bot
asgiref==3.7.2
async-timeout==4.0.2
billiard==4.1.0
cachetools==4.2.2  # Понизили до версии, совместимой с python-telegram-bot 13.13
celery==5.3.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2  # Используйте безопасную версию
defusedxml==0.7.1
Django==4.2.1
django-cors-headers==4.0.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-yasg==1.21.7
exceptiongroup==1.1.2
flake8==6.0.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.1
MarkupSafe==2.1.3
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.1
pluggy==1.2.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.8.0
pytest==7.3.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==13.13
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2